+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/dtos/gauge-chart-configuration.dto.ts
@@ -17,13 +17,13 @@ import { GraphQLJSON } from 'graphql-type-json';
 import { CalendarStartDay } from 'twenty-shared/constants';
 import {
   AggregateOperations,
+  type ChartFilter,
   type GaugeChartConfiguration,
   SerializedRelation,
 } from 'twenty-shared/types';
 
 import { UUIDScalarType } from 'src/engine/api/graphql/workspace-schema-builder/graphql-types/scalars';
 import { WidgetConfigurationType } from 'src/engine/metadata-modules/page-layout-widget/enums/widget-configuration-type.type';
-import { ChartFilter } from 'src/engine/metadata-modules/page-layout-widget/types/chart-filter.type';
 
 @ObjectType('GaugeChartConfiguration')
 export class GaugeChartConfigurationDTO implements GaugeChartConfiguration {
