+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/dtos/line-chart-configuration.dto.ts
@@ -19,6 +19,7 @@ import { GraphQLJSON } from 'graphql-type-json';
 import { CalendarStartDay } from 'twenty-shared/constants';
 import {
   AggregateOperations,
+  type ChartFilter,
   type LineChartConfiguration,
   SerializedRelation,
 } from 'twenty-shared/types';
@@ -28,7 +29,6 @@ import { AxisNameDisplay } from 'src/engine/metadata-modules/page-layout-widget/
 import { ObjectRecordGroupByDateGranularity } from 'src/engine/metadata-modules/page-layout-widget/enums/date-granularity.enum';
 import { GraphOrderBy } from 'src/engine/metadata-modules/page-layout-widget/enums/graph-order-by.enum';
 import { WidgetConfigurationType } from 'src/engine/metadata-modules/page-layout-widget/enums/widget-configuration-type.type';
-import { ChartFilter } from 'src/engine/metadata-modules/page-layout-widget/types/chart-filter.type';
 
 @ObjectType('LineChartConfiguration')
 export class LineChartConfigurationDTO implements LineChartConfiguration {
