+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/types/chart-filter.type.ts
@@ -1,20 +1,5 @@
-export type ChartRecordFilter = {
-  id: string;
-  fieldMetadataId: string;
-  operand: string;
-  value?: string | null;
-  type?: string;
-  recordFilterGroupId?: string | null;
-  subFieldName?: string | null;
-};
-
-export type ChartRecordFilterGroup = {
-  id: string;
-  logicalOperator: string;
-  parentRecordFilterGroupId?: string | null;
-};
-
-export type ChartFilter = {
-  recordFilters?: ChartRecordFilter[];
-  recordFilterGroups?: ChartRecordFilterGroup[];
-};
+export type {
+  ChartFilter,
+  ChartRecordFilter,
+  ChartRecordFilterGroup,
+} from 'twenty-shared/types';
