+++ packages/twenty-server/src/engine/metadata-modules/page-layout-widget/types/grid-position.type.ts
@@ -1,6 +1,1 @@
-export type GridPosition = {
-  row: number;
-  column: number;
-  rowSpan: number;
-  columnSpan: number;
-};
+export type { GridPosition } from 'twenty-shared/types';
