+++ packages/twenty-shared/src/types/page-layout/chart-filter.type.ts
@@ -1,0 +1,20 @@
+export type ChartRecordFilter = {
+  id: string;
+  fieldMetadataId: string;
+  operand: string;
+  value?: string | null;
+  type?: string;
+  recordFilterGroupId?: string | null;
+  subFieldName?: string | null;
+};
+
+export type ChartRecordFilterGroup = {
+  id: string;
+  logicalOperator: string;
+  parentRecordFilterGroupId?: string | null;
+};
+
+export type ChartFilter = {
+  recordFilters?: ChartRecordFilter[];
+  recordFilterGroups?: ChartRecordFilterGroup[];
+};
