+++ packages/twenty-shared/src/types/page-layout/grid-position.type.ts
@@ -1,0 +1,6 @@
+export type GridPosition = {
+  row: number;
+  column: number;
+  rowSpan: number;
+  columnSpan: number;
+};
