+++ packages/twenty-shared/src/types/page-layout/ratio-aggregate-config.type.ts
@@ -1,0 +1,6 @@
+import { type SerializedRelation } from '../SerializedRelation.type';
+
+export type RatioAggregateConfig = {
+  fieldMetadataId: SerializedRelation;
+  optionValue: string;
+};
